Support for shift on stack
PF device cannot be accessed in case of worker nodes
on OSP (as those are VMs). Use the trex server logs,
which will provide the port details. Parse the logs to
get the mac address to populate the cnfappmac. This
works for both VM and baremetal workers. And also
avoid additional job to fetch mac from PF.